Rename tool folder
