+++ tests/polymind/core_tools/test_llm_tools.py
@@ -1,6 +1,6 @@
 """Tests for OpenAIChatTool.
 Run the test with the following command:
-    poetry run pytest tests/polymind/tools/test_llm_tools.py
+    poetry run pytest tests/polymind/core_tools/test_llm_tools.py
 """
 
 import json
@@ -11,8 +11,8 @@ import pytest
 from aioresponses import aioresponses
 
 from polymind.core.message import Message
-from polymind.tools.llm_tool import OpenAIChatTool, OpenAIEmbeddingTool
-from polymind.tools.rest_api_tool import RestAPITool
+from polymind.core_tools.llm_tool import OpenAIChatTool, OpenAIEmbeddingTool
+from polymind.core_tools.rest_api_tool import RestAPITool
 
 
 class TestOpenAIChatTool:
